Split internal/db/db.go into per-domain files
Pure file moves within package db, no API changes. db.go keeps the DB
type, New/Close, migrations, and GetDBPath; each domain (users, invites,
library, activity, uploads, jobs, wishlist, requests, notifications,
webhooks, reading history, series, quality/release profiles, blocklist,
tags, monitored authors) now lives in its own file.
